Delt opp rapport-viewet i komponenter
