+++ web/src/main/vite-project/src/features/report/ErrorDetailsTable.tsx
@@ -1,0 +1,39 @@
+import ErrorLevel from "./ErrorLevel";
+import {FileReportEntryVm} from "../../kostratypes/fileReportEntryVm";
+
+
+export const ErrorDetailsTable = ({reportEntries}: { reportEntries: FileReportEntryVm[] }) =>
+    <div className="card mt-3">
+        <div className="card-body">
+            <h5 className="card-title mb-0">Feilliste</h5>
+        </div>
+        <ul className="list-group list-group-flush">
+            <li className="list-group-item">
+                <div className="table-responsive">
+                    <table className="table table-striped table-sm">
+                        <thead>
+                        <tr>
+                            <th scope="col">Feilkode</th>
+                            <th scope="col">Journalnummer</th>
+                            <th scope="col">Saksbehandler</th>
+                            <th scope="col">Kontrolltype</th>
+                            <th scope="col">Kontrolltekst</th>
+                        </tr>
+                        </thead>
+                        <tbody>
+                        {reportEntries.map((reportEntry, index) =>
+                            <tr key={index}>
+                                <th scope="row"><ErrorLevel level={reportEntry.feilkode}/></th>
+                                <td>{reportEntry.journalnummer}</td>
+                                <td>{reportEntry.saksbehandler}</td>
+                                <td>{reportEntry.kontrollnummer}</td>
+                                <td>{reportEntry.kontrolltekst}</td>
+                            </tr>)}
+                        </tbody>
+                    </table>
+                </div>
+            </li>
+        </ul>
+    </div>
+
+export default ErrorDetailsTable
